Fix saving dro change by user, re-factor dro email implementation
